[Utils cleanup] Move Kronecker utilities to the utils module
