+++ backpack/utils/kroneckers.py
